+++ packages/test-core/src/TestCore/E2EPrelude.hs
@@ -1,0 +1,3 @@
+module TestCore.E2EPrelude (module X) where
+
+import TestCore.Prelude as X