+++ packages/test-core/test/Integration/SpecPreprocessorSpec.hs
@@ -24,6 +24,20 @@ spec = describe "main" $ do
       outputContents <- readFile outputFile
       _ <- evaluate (length outputContents)
       outputContents `shouldContain'` exampleModuleHeader
+      outputContents `shouldContain'` "import TestCore.Prelude"
+
+    it "processes a simple e2e spec file" $ \(tempDir, tempFile) -> do
+      writeFile tempFile "{-# E2E_SPEC #-}"
+      let hsSourceDir = takeFileName tempDir
+          outputFile = specOutputFile tempFile
+      (exitCode, stdout, stderr) <- readProcessWithExitCode "spec-preprocessor" ["hs-source-dir=" ++ hsSourceDir, tempFile, outputFile] ""
+      stdout `shouldBe` ""
+      stderr `shouldBe` ""
+      exitCode `shouldBe` ExitSuccess
+      outputContents <- readFile outputFile
+      _ <- evaluate (length outputContents)
+      outputContents `shouldContain'` exampleModuleHeader
+      outputContents `shouldContain'` "import TestCore.E2EPrelude"
   where
     exampleModuleSegments = ["Nested"]
     exampleModuleBase = "ExampleSpec"